# docs/swagger.yaml
definitions:
  app.DumpParam:
    properties:
      dir:
        description: 冻结文件保存地址
        type: string
      leaveRunning:
        description: 在不停止程序时冻结当前程序进程
        type: boolean
      pid:
        description: 进程ID
        type: integer
      shellJob:
        description: 是否是一个命令行程序
        type: boolean
      tcpClose:
        description: 关闭已建立的TCP
        type: boolean
      tcpEstablished:
        description: 是否已建立TCP连接
        type: boolean
      tcpSkipInFlight:
        description: 是否跳过正在运行的TCP连接
        type: boolean
    type: object
  app.RestoreParam:
    properties:
      dir:
        description: 恢复已冻结进程路径
        type: string
      leaveRunning:
        description: 在不停止程序时冻结当前程序进程
        type: boolean
      rstSibling:
        description: rstSibling
        type: boolean
      shellJob:
        description: 是否是一个命令行程序
        type: boolean
      tcpClose:
        description: 关闭已建立的TCP
        type: boolean
      tcpEstablished:
        description: 是否已建立TCP连接
        type: boolean
      tcpSkipInFlight:
        description: 是否跳过正在运行的TCP连接
        type: boolean
    type: object
  app.StartCmdModel:
    properties:
      parameter:
        description: 运行所需参数
        type: string
      path:
        description: 程序所在路径
        type: string
    type: object
  response.Response:
    properties:
      code:
        description: 响应状态
        type: integer
      data:
        description: 返回数据
        type: object
      msg:
        description: 响应消息
        type: string
    type: object
info:
  contact: {}
  description: swagger Server api
  title: 仿真热迁移系统
  version: "1.0"
paths:
  /dump:
    post:
      consumes:
      - application/json
      description: 获取JSON
      parameters:
      - description: 冻结所需参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/app.DumpParam'
      responses:
        "200":
          description: 响应参数
          schema:
            $ref: '#/definitions/response.Response'
      summary: 冻结程序
      tags:
      - 冻结程序
  /restore:
    post:
      consumes:
      - application/json
      description: 获取JSON
      parameters:
      - description: 恢复所需参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/app.RestoreParam'
      responses:
        "200":
          description: 响应参数
          schema:
            $ref: '#/definitions/response.Response'
      summary: 恢复已冻结程序
      tags:
      - 恢复已冻结程序
  /startCmd:
    post:
      consumes:
      - application/json
      description: 获取JSON
      parameters:
      - description: 启动所需参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/app.StartCmdModel'
      responses:
        "200":
          description: 响应参数
          schema:
            $ref: '#/definitions/response.Response'
      summary: 根据命令参数启动程序
      tags:
      - 根据命令参数启动程序
swagger: "2.0"